requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cached-property==1.5.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
configparser==6.0.0
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
magic-filter==1.0.10
multidict==6.0.4
parsedatetime==2.6
pathlib==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
urllib3==2.0.4
yarl==1.9.2
zope.interface==6.0
